Hook the task after compilations and before test, jar and run phases
