+++ html4j-maven-plugin/src/main/resources/META-INF/gradle-plugins/html4j.properties
